trace_c4/pyproject.toml
[project]
name = "trace_c4"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ilse", email = "125893237+iklinkhamer@users.noreply.github.com"},
]
dependencies = ["npyx @ git+https://github.com/m-beau/NeuroPyxels@cell_types_classifier", "laplace-torch>=0.2.2.2", "requests>=2.32.3", "dill>=0.3.9", "tabulate>=0.9.0", "ipython>=8.31.0", "numpy>=1.26.4", "scipy==1.10.0", "pandas>=2.2.3", "numba>=0.61.0", "statsmodels>=0.14.4", "matplotlib>=3.10.0", "cmcrameri>=1.9", "scikit-learn>=1.6.1", "imbalanced-learn>=0.12.4", "networkx>=3.4.2", "psutil>=6.1.1", "joblib>=1.4.2", "tqdm>=4.67.1", "h5py>=3.12.1", "seaborn>=0.13.2", "cachecache>=0.2.2", "spikeinterface[full]>=0.101.2", "open-ephys-python-tools @ git+https://github.com/open-ephys/open-ephys-python-tools@archive"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "spyder-kernels>=3.0.2",
]
